Fixup notifier not releasing messages
After commit #83e2c5a4 whenever a message was recieved, it wass queued as normally
but after a message was released from the queue the next message was not released
before a new message was incomming to the socket.

This made the queue fail when there where multiple messages queued up as they would never be released
and on next recieve of a message a old one in the queue was released making it quite useless.

This fixes the issue by processing the queue for every execution of poll()
running poll() until its empty.