Decouple event, log, history from wire protocol
The EventEngine currently only supports the first load of events, not
updates.

This was originally written with pydantic v1, v2 was released during
development of this rather long-lived branch. Since v2 made our subclass
of BaseModel unnecessary, we attempted to upgrade via bump-pydantic. It
didn't pick up on our sub-class and some other annoyances.

Input to timedelta works different, on v1 you could build a timedelta
from an int of seconds stored in a string, in v2 it must be an int.
Interestingly the v1 docs does not mention that int-as-string is
supported.