Make it work on pydantic 2
It turns out that bump-pydantic only does so much. It didn't aattempt to
fix subclasses of BaseModel, fair.

Input to timedelta works different, on v1 you could build a timedelta
from an int of seconds stored in a string, in v2 it must be an int.
Interestingly the v1 docs does not mention that int-as-string is
supported.